fixes no dependncy config in readme so app can be run with no security if required
