+++ app/spring/src/main/java/uk/co/mruoc/nac/app/config/security/SecurityConfig.java
@@ -2,7 +2,7 @@ package uk.co.mruoc.nac.app.config.security;
 
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
-import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
@@ -13,20 +13,10 @@ import org.springframework.security.oauth2.server.resource.authentication.JwtAut
 import org.springframework.security.web.SecurityFilterChain;
 import uk.co.mruoc.nac.app.config.websocket.AuthChannelInterceptor;
 import uk.co.mruoc.nac.app.config.websocket.DefaultAuthChannelInterceptor;
-import uk.co.mruoc.nac.app.security.SpringAuthenticatedUserSupplier;
-import uk.co.mruoc.nac.usecases.AuthCodeClient;
-import uk.co.mruoc.nac.usecases.AuthService;
-import uk.co.mruoc.nac.usecases.AuthenticatedUserSupplier;
-import uk.co.mruoc.nac.usecases.AuthenticatedUserValidator;
-import uk.co.mruoc.nac.usecases.ExternalUserPresentRetry;
-import uk.co.mruoc.nac.usecases.ExternalUserService;
-import uk.co.mruoc.nac.usecases.ExternalUserSynchronizer;
-import uk.co.mruoc.nac.usecases.TokenService;
-import uk.co.mruoc.nac.usecases.UserFinder;
 
 @Configuration
 @EnableMethodSecurity
-@Slf4j
+@ConditionalOnProperty(name = "auth.security.enabled", havingValue = "true", matchIfMissing = true)
 public class SecurityConfig {
 
   @Bean
@@ -49,29 +39,5 @@ public class SecurityConfig {
   @Bean
   public AuthChannelInterceptor authChannelInterceptor(JwtDecoder jwtDecoder) {
     return new DefaultAuthChannelInterceptor(jwtDecoder);
-  }
-
-  @Bean
-  public AuthenticatedUserSupplier authenticatedUserSupplier(UserFinder userFinder) {
-    return new SpringAuthenticatedUserSupplier(userFinder);
-  }
-
-  @Bean
-  public AuthenticatedUserValidator authenticatedUserValidator(AuthenticatedUserSupplier supplier) {
-    return new AuthenticatedUserValidator(supplier);
-  }
-
-  @Bean
-  public AuthService authService(
-      TokenService tokenService,
-      AuthCodeClient authCodeClient,
-      ExternalUserService externalUserService,
-      ExternalUserSynchronizer synchronizer) {
-    return AuthService.builder()
-        .tokenService(tokenService)
-        .authCodeClient(authCodeClient)
-        .externalUserPresentRetry(new ExternalUserPresentRetry(externalUserService))
-        .synchronizer(synchronizer)
-        .build();
   }
 }
